move ResultsDocValueIterator to SeismicBaseScorer
