+++ src/main/java/org/opensearch/neuralsearch/sparse/algorithm/SeismicBaseScorer.java
@@ -271,4 +271,55 @@ public abstract class SeismicBaseScorer extends Scorer {
             return 0;
         }
     }
+
+    public static class ResultsDocValueIterator extends DocIdSetIterator {
+        private final IteratorWrapper<Pair<Integer, Integer>> resultsIterator;
+        private int docId;
+
+        public ResultsDocValueIterator(List<Pair<Integer, Integer>> results) {
+            resultsIterator = new IteratorWrapper<>(results.iterator());
+            docId = -1;
+        }
+
+        @Override
+        public int docID() {
+            return docId;
+        }
+
+        @Override
+        public int nextDoc() throws IOException {
+            if (resultsIterator.next() == null) {
+                docId = NO_MORE_DOCS;
+                return NO_MORE_DOCS;
+            }
+            docId = resultsIterator.getCurrent().getLeft();
+            return docId;
+        }
+
+        @Override
+        public int advance(int target) throws IOException {
+            if (target <= docId) {
+                return docId;
+            }
+            while (resultsIterator.hasNext()) {
+                Pair<Integer, Integer> pair = resultsIterator.next();
+                if (pair.getKey() >= target) {
+                    docId = pair.getKey();
+                    return docId;
+                }
+            }
+            docId = NO_MORE_DOCS;
+            return NO_MORE_DOCS;
+        }
+
+        // we use cost() to return prestored score
+        @Override
+        public long cost() {
+            if (resultsIterator.getCurrent() == null || !resultsIterator.hasNext()) {
+                return 0;
+            } else {
+                return resultsIterator.getCurrent().getValue();
+            }
+        }
+    }
 }
